+++ tasks/options/eslint.js
@@ -1,0 +1,11 @@
+module.exports = {
+    all: {
+        src: [
+            'src/**/*.js',
+            'tests/**/*.js',
+            'Gruntfile.js',
+            'scripts/**/*.js',
+            'tasks/**/*.js'
+        ]
+    }
+};