+++ tests/e2e/cwrxStream.e2e.js
@@ -57,7 +57,7 @@ describe('cwrxStream', function() {
     beforeEach(function() {
         jasmine.DEFAULT_TIMEOUT_INTERVAL = 120000;
         var awsConfig = {
-            region: 'us-east-1',
+            region: 'us-east-1'
         };
         if(AWS_CREDS) {
             awsConfig.accessKeyId = AWS_CREDS.accessKeyId;
@@ -533,6 +533,26 @@ describe('cwrxStream', function() {
                     });
                 }).catch(done.fail);
             });
+
+            it('should indicate in Hubspot that this user is a paying customer', function(done) {
+                producer.produce({
+                    type: 'paymentMade',
+                    data: {
+                        payment: mockPayment,
+                        user: mockUser,
+                        balance: 9001.12,
+                        target: 'showcase'
+                    }
+                }).then(() => {
+                    return Promise.all([
+                        waitForTrue(() =>  this.hubspot.getContactByEmail(mockUser.email)),
+                        waitForEmails(['Your payment has been approved'])
+                    ]);
+                }).then(results => {
+                    const contact = results[0];
+                    expect(contact.properties.paying_customer.value).toBe('true');
+                }).then(done, done.fail);
+            });
         });
     });
 
@@ -687,7 +707,6 @@ describe('cwrxStream', function() {
         });
 
         describe('with a promotion', function() {
-            /* jshint camelcase: false */
             var testOrg, testPromotions;
             beforeEach(function(done) {
                 mockUser.org = 'o-e2e-1';
@@ -814,8 +833,6 @@ describe('cwrxStream', function() {
                     expect(results.rows.length).toBe(0);
                 }).then(done, done.fail);
             });
-
-            /* jshint camelcase: true */
         });
 
         describe('in hubspot', function() {
@@ -918,8 +935,7 @@ describe('cwrxStream', function() {
                     target: 'showcase'
                 }
             }).then(function() {
-                mailman.once('Reelcontent Password Change Notice',
-                        function(msg) {
+                mailman.once('Reelcontent Password Change Notice', function(msg) {
                     expect(msg.from[0].address.toLowerCase()).toBe('support@cinema6.com');
                     expect(msg.to[0].address.toLowerCase()).toBe('c6e2etester@gmail.com');
                     var regex = /Terry, Just a quick note/;