+++ tests/unit/check_signup_promotion.ut.js
@@ -17,27 +17,26 @@ function series(fns) {
 }
 
 describe('check_signup_promotion.js', function() {
-    var mockOptions, mockConfig, mockLog, mockProducer, event, mockOrg, mockPromotion,
+    var mockConfig, mockLog, mockProducer, event, mockOrg, mockPromotion,
         resps, checkSignupProm;
 
     beforeEach(function() {
         jasmine.clock().install();
         jasmine.clock().mockDate(FAKE_NOW);
     });
-    
+
     afterEach(function() {
         jasmine.clock().uninstall();
     });
 
     beforeEach(function() {
-        mockOptions = { };
         mockConfig = {
             appCreds: 'i am watchman',
             cwrx: {
                 api: {
                     root: 'http://test.com',
                     promotions: {
-                        endpoint: '/api/promotions',
+                        endpoint: '/api/promotions'
                     },
                     orgs: {
                         endpoint: '/api/account/orgs'
@@ -60,7 +59,7 @@ describe('check_signup_promotion.js', function() {
             promotion: 'pro-signup-1',
             status: 'active'
         } } };
-        
+
         mockLog = {
             trace : jasmine.createSpy('log.trace'),
             error : jasmine.createSpy('log.error'),
@@ -95,13 +94,13 @@ describe('check_signup_promotion.js', function() {
         resps = {
             orgs: {
                 get: { response: { statusCode: 200 }, body: mockOrg },
-                put: { response: { statusCode: 200 }, body: mockOrg },
+                put: { response: { statusCode: 200 }, body: mockOrg }
             },
             promotions: {
                 get: { response: { statusCode: 200 }, body: mockPromotion }
             }
         };
-        
+
         spyOn(requestUtils, 'makeSignedRequest').and.callFake(function(creds, method, opts) {
             if (/orgs/.test(opts.url)) {
                 return q(resps.orgs[method]);
@@ -109,7 +108,7 @@ describe('check_signup_promotion.js', function() {
                 return q(resps.promotions[method]);
             }
         });
-        
+
         checkSignupProm = actionFactory(mockConfig);
     });
 
@@ -169,7 +168,7 @@ describe('check_signup_promotion.js', function() {
             done();
         }).catch(done.fail);
     });
-    
+
     it('should initialize the promotions array if not defined on the org', function(done) {
         var expectedArr = [{
             id: 'pro-signup-1',
@@ -196,7 +195,7 @@ describe('check_signup_promotion.js', function() {
             done();
         }).catch(done.fail);
     });
-    
+
     it('should warn and skip if the promotion is invalid', function(done) {
         mockPromotion.status = Status.Inactive;
         checkSignupProm(event).then(function() {
@@ -204,7 +203,7 @@ describe('check_signup_promotion.js', function() {
             expect(requestUtils.makeSignedRequest)
                 .not.toHaveBeenCalledWith('i am watchman', 'put', jasmine.any(Object));
             expect(mockLog.warn).toHaveBeenCalled();
-            
+
             requestUtils.makeSignedRequest.calls.reset();
             mockLog.warn.calls.reset();
             mockPromotion.status = Status.Active;
@@ -220,7 +219,7 @@ describe('check_signup_promotion.js', function() {
             done();
         }).catch(done.fail);
     });
-    
+
     it('should warn and skip if the org already has the promotion', function(done) {
         mockOrg.promotions[0].id = event.data.user.promotion;
         checkSignupProm(event).then(function() {
@@ -260,13 +259,13 @@ describe('check_signup_promotion.js', function() {
             expect(mockProducer.produce).not.toHaveBeenCalled();
         }).then(done, done.fail);
     });
-    
+
     it('should warn and skip if fetching the promotion returns a 4xx', function(done) {
         resps.promotions.get = {
             response: { statusCode: 404 },
             body: 'dat aint real'
         };
-        
+
         checkSignupProm(event).then(function() {
             expect(requestUtils.makeSignedRequest.calls.count()).toBe(2);
             expect(requestUtils.makeSignedRequest)
@@ -277,7 +276,7 @@ describe('check_signup_promotion.js', function() {
             done();
         }).catch(done.fail);
     });
-    
+
     ['get', 'put'].forEach(function(verb) {
         var reqStr = verb.toUpperCase() + ' /orgs';
 
@@ -286,7 +285,7 @@ describe('check_signup_promotion.js', function() {
                 response: { statusCode: 400 },
                 body: 'I got a problem with YOU'
             };
-            
+
             checkSignupProm(event).then(function() {
                 if (verb === 'get') {
                     expect(requestUtils.makeSignedRequest)
@@ -304,7 +303,7 @@ describe('check_signup_promotion.js', function() {
             }).catch(done.fail);
         });
     });
-    
+
     [{ obj: 'orgs', verb: 'get' }, { obj: 'orgs', verb: 'put' }, { obj: 'promotions', verb: 'get' }]
     .forEach(function(cfg) {
         var reqStr = cfg.verb.toUpperCase() + ' /' + cfg.obj;